Consolidation of utils and duplicated code.
