+++ container/conductor/engine.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .visibility import getLogger
+from container.utils.visibility import getLogger
 logger = getLogger(__name__)
 
 CAPABILITIES = dict(
